commented out some code

+++ Deadwood/src/deadwood/BoardPane.java
@@ -131,7 +131,7 @@ public class BoardPane extends JLayeredPane {
     public void movePlayerLabel(int playerID, int x, int y) {
         //System.out.println(x);
         playerLabels[playerID].setBounds(x, y, playerIconWidth, playerIconHeight);
-        add(playerLabels[playerID], new Integer(2));
+        add(playerLabels[playerID], new Integer(3));
 //        InfoPanel.getInstance().setPlayerInfoData(active);
 //        InfoPanel.getInstance().setUpdateTextArea("Your have moved !\n");
     }
